refactor: 동아리 관리자 도메인을 하위 패키지로 이동
