+++ src/main/java/in/koreatech/koin/admin/club/service/AdminClubService.java
@@ -15,10 +15,10 @@ import in.koreatech.koin.domain.club.category.model.ClubCategory;
 import in.koreatech.koin.domain.club.category.repository.ClubCategoryRepository;
 import in.koreatech.koin.domain.club.club.exception.ClubNotFoundException;
 import in.koreatech.koin.domain.club.club.model.Club;
-import in.koreatech.koin.domain.club.club.model.ClubManager;
+import in.koreatech.koin.domain.club.manager.model.ClubManager;
 import in.koreatech.koin.domain.club.club.model.ClubSNS;
 import in.koreatech.koin.domain.club.club.model.redis.ClubCreateRedis;
-import in.koreatech.koin.domain.club.club.repository.ClubManagerRepository;
+import in.koreatech.koin.domain.club.manager.repository.ClubManagerRepository;
 import in.koreatech.koin.domain.club.club.repository.ClubRepository;
 import in.koreatech.koin.domain.club.club.repository.redis.ClubCreateRedisRepository;
 import in.koreatech.koin.domain.user.model.User;
