+++ src/main/java/in/koreatech/koin/domain/club/club/model/redis/ClubCreateRedis.java
@@ -12,7 +12,7 @@ import org.springframework.data.redis.core.TimeToLive;
 import in.koreatech.koin.domain.club.club.dto.request.ClubCreateRequest;
 import in.koreatech.koin.domain.club.club.model.Club;
 import in.koreatech.koin.domain.club.category.model.ClubCategory;
-import in.koreatech.koin.domain.club.club.model.ClubManager;
+import in.koreatech.koin.domain.club.manager.model.ClubManager;
 import in.koreatech.koin.domain.user.model.User;
 import jakarta.persistence.Id;
 import lombok.Builder;
