+++ src/main/java/in/koreatech/koin/domain/club/event/service/ClubEventService.java
@@ -1,6 +1,6 @@
 package in.koreatech.koin.domain.club.event.service;
 
-import in.koreatech.koin.domain.club.club.repository.ClubManagerRepository;
+import in.koreatech.koin.domain.club.manager.repository.ClubManagerRepository;
 import in.koreatech.koin.domain.club.club.repository.ClubRepository;
 import in.koreatech.koin.domain.club.event.dto.request.ClubEventCreateRequest;
 import in.koreatech.koin.domain.club.event.dto.request.ClubEventModifyRequest;
@@ -15,6 +15,7 @@ import in.koreatech.koin.domain.club.club.model.Club;
 import in.koreatech.koin.domain.club.event.model.ClubEvent;
 import in.koreatech.koin.domain.club.event.model.ClubEventImage;
 import in.koreatech.koin.domain.club.event.model.ClubEventSubscription;
+import in.koreatech.koin.domain.club.manager.service.ClubManagerService;
 import in.koreatech.koin.domain.student.model.Student;
 import in.koreatech.koin.domain.student.repository.StudentRepository;
 import in.koreatech.koin.domain.user.model.User;
@@ -43,6 +44,7 @@ public class ClubEventService {
     private final ClubEventSubscriptionRepository clubEventSubscriptionRepository;
     private final ClubManagerRepository clubManagerRepository;
     private final UserRepository userRepository;
+    private final ClubManagerService clubManagerService;
 
     @Transactional
     public void createClubEvent(ClubEventCreateRequest request, Integer clubId, Integer studentId) {
