+++ src/main/java/in/koreatech/koin/domain/club/manager/dto/request/ClubManagerEmpowermentRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.club.club.dto.request;
+package in.koreatech.koin.domain.club.manager.dto.request;
 
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
