+++ src/main/java/in/koreatech/koin/domain/club/manager/model/ClubManager.java
@@ -1,9 +1,10 @@
-package in.koreatech.koin.domain.club.club.model;
+package in.koreatech.koin.domain.club.manager.model;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
+import in.koreatech.koin.domain.club.club.model.Club;
 import in.koreatech.koin.domain.user.model.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
